requirements.txt
actionlib==1.14.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
angles==1.9.13
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
bondpy==1.8.6
cffi==1.16.0
comm==0.2.0
commentjson==0.9.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dill==0.3.7
distro==1.8.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
empy==3.3.4
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
fonttools==4.46.0
fqdn==1.5.1
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
image-geometry==1.16.2
importlib-resources==6.1.1
interactive-markers==1.12.0
ipykernel==6.27.1
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_packaging==0.12.3
jupyter_server==2.11.1
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
lark-parser==0.7.8
laser_geometry==1.6.7
lazy_loader==0.3
lightning-utilities==0.10.0
llvmlite==0.37.0
matplotlib-inline==0.1.6
message-filters==1.16.0
mistune==3.0.2
msgpack==1.0.7
multiprocess==0.70.15
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.24.4
open3d==0.15.2
opencv-contrib-python==4.8.1.78
overrides==7.4.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
plyfile==1.0.2
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyquaternion==0.9.9
pyrealsense2==2.54.1.5217
python-json-logger==2.0.7
python-qt-binding==0.4.4
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.31.0
resource_retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.13.1
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.20
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sensor-msgs==1.13.1
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
topic-tools==1.16.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==1.2.1
torchvision==0.14.1+cu117
traitlets==5.14.0
trimesh==4.0.5
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xacro==1.14.17
y-py==0.6.2
ypy-websocket==0.8.4